properly poll builds, document API
This makes use of headers Jenkins gives back to us to indicate where we
should be next in the polling of a job, and when that job is done.

This also makes some additional effort to document the Jenkins API,
since it doesn't seem much of this documentation is available elsewhere.